Add user authorization and PrivateRoute for most pages
1. I added a new component: PrivateRoute that gets called
when user is trying to access user-id dependent routes:
/new, /workouts, /personal-bests.
2. I created a Login component that is displayed to the user
when they are trying to access a route they need to be
registered to view. There, they can log in/sign up.
If the user is already logged in and somehow went to the /Login
page, it says "you're already signed in".
3. In RouteSwitch, I wrapped the private routes in PrivateRoute
that checks whether user is logged in. If yes, it redirects
them to that page. If not, it directs the user to /login.